reworked how the API returns data
Changed all routes to return better statuscodes and to return json data with "res.json" instead of doing alot of "parse.JSON".
Changed parameters for some requests to in path. this was because RFC doesnt mention anything about using a body on delete requests-
There is also a in path parameter on the insertall in the eadmin route. This was just to code less.

mongo.js doesnt have alot of errorhandling, the one portion where there will be an error on a duplicate insert, will now just return a text "errored" and the controller will do soemthing with that.